chore: reorganizing TaskRestServiceTest.kt
