+++ src/test/kotlin/nl/info/zac/app/documentcreation/DocumentCreationRestServiceTest.kt
@@ -13,13 +13,13 @@ import io.mockk.every
 import io.mockk.mockk
 import io.mockk.slot
 import net.atos.zac.admin.ZaakafhandelParameterService
-import net.atos.zac.flowable.createTestTask
 import net.atos.zac.flowable.task.FlowableTaskService
 import net.atos.zac.flowable.task.exception.TaskNotFoundException
 import nl.info.client.zgw.model.createZaak
 import nl.info.client.zgw.zrc.ZrcClientService
 import nl.info.client.zgw.ztc.ZtcClientService
 import nl.info.client.zgw.ztc.model.createInformatieObjectType
+import nl.info.test.org.flowable.task.api.createTestTask
 import nl.info.zac.admin.model.createZaakafhandelParameters
 import nl.info.zac.app.documentcreation.model.createRestDocumentCreationAttendedData
 import nl.info.zac.documentcreation.BpmnDocumentCreationService
